update kernel video/dsscomp.h
dsscomp: remove name collision for DISPLAY_SETUP_DISPLAY

This was defined both as an enum (flag) and as a define. It is
now separated in dsscomp.h

A patch for this was already merged directly in bionic:
https://android.dal.design.ti.com/#change,5583
This patch is just a sync so we don't revert the above one if new
changes are done here in dsscomp.h header.

Change-Id: Ibaf34953ee8bd3e617735e818c99507f6daf6053
Signed-off-by: Pavel Nedev <pnedev@mm-sol.com>